requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_kmp_llvm
aom=3.5.0=h27087fc_0
bzip2=1.0.8=h7f98852_4
ca-certificates=2022.12.7=ha878542_0
cffi=1.14.6=py36hd8eec40_1
cloudpickle=1.6.0=py_0
cycler=0.11.0=pypi_0
dataclasses=0.8=pyh787bdff_2
expat=2.5.0=h27087fc_0
ffmpeg=5.1.2=gpl_h8dda1f0_106
font-ttf-dejavu-sans-mono=2.37=hab24e00_0
font-ttf-inconsolata=3.000=h77eed37_0
font-ttf-source-code-pro=2.038=h77eed37_0
font-ttf-ubuntu=0.83=hab24e00_0
fontconfig=2.14.2=h14ed4e7_0
fonts-conda-ecosystem=1=0
fonts-conda-forge=1=0
freetype=2.12.1=hca18f0e_1
future=0.18.2=py36h5fab9bb_3
gettext=0.21.1=h27087fc_0
gmp=6.2.1=h58526e2_0
gnutls=3.7.8=hf3e180e_0
gym=0.19.0=py36he75451f_0
gym-notices=0.0.8=pyhd8ed1ab_0
icu=70.1=h27087fc_0
kiwisolver=1.3.1=pypi_0
lame=3.100=h166bdaf_1003
ld_impl_linux-64=2.40=h41732ed_0
libblas=3.9.0=12_linux64_mkl
libcblas=3.9.0=12_linux64_mkl
libdrm=2.4.114=h166bdaf_0
libffi=3.4.2=h7f98852_5
libgcc-ng=12.2.0=h65d4601_19
libgfortran-ng=12.2.0=h69a702a_19
libgfortran5=12.2.0=h337968e_19
libgomp=12.2.0=h65d4601_19
libhwloc=2.9.0=hd6dc26d_0
libiconv=1.17=h166bdaf_0
libidn2=2.3.4=h166bdaf_0
liblapack=3.9.0=12_linux64_mkl
libnsl=2.0.0=h7f98852_0
libopenblas=0.3.21=pthreads_h78a6416_3
libopus=1.3.1=h7f98852_1
libpciaccess=0.17=h166bdaf_0
libpng=1.6.39=h753d276_0
libprotobuf=3.18.3=h6239696_0
libsqlite=3.40.0=h753d276_0
libstdcxx-ng=12.2.0=h46fd767_19
libtasn1=4.19.0=h166bdaf_0
libunistring=0.9.10=h7f98852_0
libuuid=2.32.1=h7f98852_1000
libva=2.17.0=h0b41bf4_0
libvpx=1.11.0=h9c3ff4c_3
libxcb=1.13=h7f98852_1004
libxml2=2.10.3=h7463322_0
libzlib=1.2.13=h166bdaf_4
llvm-openmp=15.0.7=h0cdce71_0
matplotlib=3.3.4=pypi_0
mkl=2021.4.0=h8d4b97c_729
ncurses=5.9=10
nettle=3.8.1=hc379101_1
ninja=1.11.1=h924138e_0
numpy=1.19.5=py36hfc0c790_2
openh264=2.3.1=hcb278e6_2
openssl=1.0.2u=h516909a_0
p11-kit=0.24.1=hc5aa10d_0
pillow=8.4.0=pypi_0
pip=21.3.1=pyhd8ed1ab_0
protobuf=3.3.2=py36_0
pthread-stubs=0.4=h36c2ea0_1001
pycparser=2.21=pyhd8ed1ab_0
pygame=2.1.3=pypi_0
pyglet=1.5.16=py36h5fab9bb_0
pyparsing=3.0.7=pypi_0
python=3.6.1=3
python-dateutil=2.8.2=pypi_0
python_abi=3.6=2_cp36m
pytorch=1.9.1=cpu_py36ha8b20dc_3
readline=6.2=0
scipy=1.5.3=py36h81d768a_1
setuptools=58.0.4=py36h5fab9bb_2
six=1.16.0=pyh6c4a22f_0
sleef=3.5.1=h9b69904_2
sqlite=3.13.0=1
svt-av1=1.4.1=hcb278e6_0
tbb=2021.8.0=hf52228f_0
tensorboardx=2.4=pyhd8ed1ab_0
tk=8.5.19=2
typing_extensions=4.1.1=pyha770c72_0
tzdata=2022g=h191b570_0
wheel=0.37.1=pyhd8ed1ab_0
x264=1!164.3095=h166bdaf_2
x265=3.5=h924138e_3
xorg-fixesproto=5.0=h7f98852_1002
xorg-kbproto=1.0.7=h7f98852_1002
xorg-libx11=1.7.2=h7f98852_0
xorg-libxau=1.0.9=h7f98852_0
xorg-libxdmcp=1.1.3=h7f98852_0
xorg-libxext=1.3.4=h0b41bf4_2
xorg-libxfixes=5.0.3=h7f98852_1004
xorg-xextproto=7.3.0=h0b41bf4_1003
xorg-xproto=7.0.31=h7f98852_1007
xz=5.2.6=h166bdaf_0
zlib=1.2.13=h166bdaf_4
